+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## Change Log
 
 ### [Unreleased][unreleased]
-### Added
+#### Added
+- Multi write connections support
 - Message serializer change support
 
 ### [1.5.0] - 2025-04-01